alarm-rag-agent/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.6
streamlit>=1.34
pandas>=2.1
scikit-learn>=1.3
python-dotenv>=1.0
langchain>=0.3
langchain-community>=0.3
langchain-core>=0.3
langchain-groq>=0.2
langchain-huggingface>=0.1
langgraph>=0.2
sentence-transformers>=2.7
faiss-cpu>=1.8
rank_bm25>=0.2.2
pytest>=8.0
httpx>=0.27
